core_new/organizations/backend/ecs_tasks/deploy_cdk/cdkproxy/requirements.txt
aws-cdk-lib==2.20.0
aws_cdk.aws_redshift_alpha==2.14.0a0
boto3==1.24.85
boto3-stubs==1.24.85
botocore==1.27.85
cdk-nag==2.7.2
constructs==10.0.73
fastapi==0.78.0
Flask==2.0.2
PyYAML==6.0
requests==2.27.1
tabulate==0.8.9
uvicorn==0.15.0
jinja2==3.0.3
werkzeug==2.0.2
constructs>=10.0.0,<11.0.0
git-remote-codecommit==1.16
aws-ddk==0.5.1
aws-ddk-core==0.5.1